🎨 Refacto actions candidature
